Update anydesk.tar.gz to 5.1.2
The generic Linux links on the front page is actually broken. I
discovered that https://download.anydesk.com/linux/ has an index of all
past revisions of the application, and 5.1.2 is present there. They
appear to have switched from "i686" to "i386" in this release, too. I
hacked the checker locally to insert "/generic-linux" into the URL after
scraping the download page.

I'm not sure how best to handle this; some options might be:

* Approach Anydesk and encourage them to:
    a) fix the broken generic Linux links on
       https://anydesk.com/en/downloads/linux
    b) link to these stable URLs that don't disappear when they release
       a new version
* Teach the checker to scrape this page and pick the matching URL with
  the highest version number
* Switch this Flatpak over to using the .debs, at which point it can use
  the apt repository to find the latest version

Fixes #12.
Fixes #15.