Don't try to preserve Anydesk archive content file owners
When extracting as root (i.e. when flatpak was run with root privileges),
tar defaults to --same-owner which leads to lots of ownership change
errors like the following:

    tar: anydesk: Cannot change ownership to uid 109, gid 114: Operation
      not permitted

This happens because files inside the archive have 109 and 114 as their
owner and group, respectively.

This commit adds --no-same-owner tar option to get rid of these errors.